# docs/specs/openapi.yaml
openapi: "3.1.0"
info:
  title: Aether Platform API
  description: |
    White-label SaaS коммуникационная платформа.
    Multi-tenant, plugin architecture, AI-powered.

    ## Аутентификация
    - **Bearer JWT**: `Authorization: Bearer <access_token>`
    - **API Key**: `X-API-Key: ak_...`

    ## Tenant isolation
    Все tenant-specific эндпоинты требуют JWT с `organisation_id` claim.
    Admin эндпоинты требуют `is_superadmin = true`.

    ## Rate Limiting
    - Auth: IP-based (5 login/min, 3 signup/hour)
    - Tenant: per-plan (60-6000 RPM)
    - Headers: `X-RateLimit-Limit`, `X-RateLimit-Remaining`, `X-RateLimit-Reset`

    ## Pagination
    Cursor-based: `?cursor=<last_id>&limit=<1-100>`, default 20.
    Response: `{"items": [...], "next_cursor": "uuid"}`.

    ## Errors
    Все ошибки возвращают JSON: `{"error": {"code": "...", "message": "...", "details": {...}}}`
  version: "1.0.0"
  contact:
    name: Aether Platform
    email: dev@aether.cloud
servers:
  - url: https://api.aether.cloud/api/v1
    description: Production
  - url: http://localhost:8000/api/v1
    description: Development
security:
  - bearerAuth: []
  - apiKeyAuth: []

# ============================================================================
# SECURITY SCHEMES
# ============================================================================
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token from /auth/login or /auth/refresh
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: 'API key from /auth/api-keys (prefix: ak_)'

  # ============================================================================
  # COMMON SCHEMAS
  # ============================================================================
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Machine-readable error code
              example: "invalid_credentials"
            message:
              type: string
              description: Human-readable error message (localized)
              example: "Неверный email или пароль"
            details:
              type: object
              description: Field-level validation errors
    ValidationError:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, details]
          properties:
            code:
              type: string
              example: "validation_error"
            message:
              type: string
              example: "Ошибка валидации"
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string

    # --- Pagination ---
    PaginatedResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: {}
        next_cursor:
          type: string
          format: uuid
          nullable: true
          description: Cursor for next page, null if last page
        total:
          type: integer
          description: Total count (when available)
          example: 42

    # --- Auth ---
    SignupRequest:
      type: object
      required: [email, company_name]
      properties:
        email:
          type: string
          format: email
          minLength: 5
          maxLength: 254
          description: Email пользователя
          example: "ivan@logistics.ru"
        company_name:
          type: string
          minLength: 2
          maxLength: 100
          description: Название компании
          example: "Логистика Плюс"
        locale:
          type: string
          default: "ru"
          enum: [ru, en]
          description: Язык интерфейса
    LoginRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
          description: Email для входа
          example: "ivan@logistics.ru"
    LoginChallenge:
      type: object
      description: Ответ на /auth/login — какие методы доступны
      properties:
        has_passkey:
          type: boolean
          description: Есть ли зарегистрированный passkey
        has_password:
          type: boolean
          description: Установлен ли пароль
        has_mfa:
          type: boolean
          description: Включена ли MFA
        magic_link_sent:
          type: boolean
          description: Отправлена ли magic link (если нет других методов)
        passkey_options:
          type: object
          nullable: true
          description: WebAuthn PublicKeyCredentialRequestOptions
        available_providers:
          type: array
          items:
            type: string
            enum: [google, yandex, vk]
          description: Доступные OAuth провайдеры
...
    MagicLinkRequest:
      type: object
      required: [token]
      properties:
        token:
          type: string
          minLength: 32
          description: Token из magic link URL
...
    TokenResponse:
      type: object
      required: [access_token, refresh_token, expires_in, token_type, user]
      properties:
        access_token:
          type: string
          description: JWT access token
        refresh_token:
          type: string
          description: JWT refresh token
        expires_in:
          type: integer
          description: TTL access token в секундах
          example: 3600
        token_type:
          type: string
          example: "bearer"
        user:
          $ref: '#/components/schemas/UserResponse'
        workspace:
          $ref: '#/components/schemas/WorkspaceInfo'
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
          description: JWT refresh token
    WorkspaceInfo:
      type: object
      required: [organisation_id, slug, name]
      properties:
        organisation_id:
          type: string
          format: uuid
        slug:
          type: string
          description: URL-safe идентификатор
        name:
          type: string
          description: Название компании
        logo_url:
          type: string
          format: uri
          nullable: true
        role:
          type: string
          enum: [owner, admin, member, viewer]
        available_channels:
          type: array
          items:
            type: string
          description: Доступные типы каналов

    # --- Users ---
    UserResponse:
      type: object
      required: [id, email, full_name, is_active, created_at]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        full_name:
          type: string
        avatar_url:
          type: string
          format: uri
          nullable: true
        is_active:
          type: boolean
        is_superadmin:
          type: boolean
        mfa_enabled:
          type: boolean
        last_login_at:
          type: string
          format: date-time
          nullable: true
        locale:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    ApiKeyResponse:
      type: object
      required: [id, name, key_prefix, scopes, created_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        key_prefix:
          type: string
          description: Первые 8 символов для идентификации
          example: "ak_a1b2c3d4"
        scopes:
          type: array
          items: { type: string }
        last_used_at:
          type: string
          format: date-time
          nullable: true
        expires_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
    ApiKeyCreate:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        scopes:
          type: array
          items: { type: string }
          default: ["*"]
        expires_at:
          type: string
          format: date-time
    ApiKeyCreated:
      type: object
      required: [id, name, key, key_prefix]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        key:
          type: string
          description: Полный API ключ — показывается ТОЛЬКО при создании
          example: "ak_a1b2c3d4e5f6g7h8i9j0k1l2m3n4o5p6q7r8s9t0"
        key_prefix:
          type: string

    # --- Audit ---
...
  parameters:
    CursorParam:
      name: cursor
      in: query
      schema:
        type: string
        format: uuid
      description: Cursor для пагинации (id последнего элемента предыдущей страницы)
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Количество элементов на странице

  # ============================================================================
  # COMMON RESPONSES
  # ============================================================================
  responses:
    401Unauthorized:
      description: Не авторизован (JWT истёк или отсутствует)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: "unauthorized"
              message: "Требуется аутентификация"
...
    404NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    422ValidationError:
      description: Ошибка валидации
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
    429RateLimited:
      description: Превышен лимит запросов
      headers:
        X-RateLimit-Limit:
          schema: { type: integer }
        X-RateLimit-Remaining:
          schema: { type: integer }
        X-RateLimit-Reset:
          schema: { type: integer }
        Retry-After:
          schema: { type: integer }
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: "rate_limited"
              message: "Слишком много запросов. Попробуйте через 30 секунд."

# ============================================================================
# PATHS
# ============================================================================
paths:
...
  /auth/signup:
    post:
      operationId: signup
      summary: Регистрация нового пользователя и tenant'а
      description: Создаёт пользователя, организацию и отправляет welcome email.
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupRequest'
      responses:
        '201':
          description: Пользователь и организация созданы
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          $ref: '#/components/responses/422ValidationError'
        '429':
          $ref: '#/components/responses/429RateLimited'
  /auth/login:
    post:
      operationId: login
      summary: Начать процесс входа
      description: Возвращает доступные методы входа для email. Если пароля/passkey нет — автоматически отправляет magic link.
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Доступные методы
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginChallenge'
        '400':
          $ref: '#/components/responses/422ValidationError'
        '429':
          $ref: '#/components/responses/429RateLimited'
...
  /auth/login/magic-link:
    post:
      operationId: loginMagicLink
      summary: Вход по magic link токену
      description: Обменивает magic link токен на JWT пару.
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MagicLinkRequest'
      responses:
        '200':
          description: Успешный вход
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '429':
          $ref: '#/components/responses/429RateLimited'
...
  /auth/refresh:
    post:
      operationId: refreshToken
      summary: Обновить access токен
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Новая пара токенов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/401Unauthorized'
  /auth/logout:
    post:
      operationId: logout
      summary: Выход (инвалидация refresh токена)
      tags: [Auth]
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Токен инвалидирован
...
  /users/me:
    get:
      operationId: getCurrentUser
      summary: Профиль текущего пользователя
      tags: [Users]
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      responses:
        '200':
          description: Профиль пользователя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/401Unauthorized'
    patch:
      operationId: updateCurrentUser
      summary: Обновить профиль
      tags: [Users]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                full_name:
                  type: string
                  minLength: 1
                  maxLength: 255
                locale:
                  type: string
                  enum: [ru, en]
                avatar_url:
                  type: string
                  format: uri
      responses:
        '200':
          description: Обновлённый профиль
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '422':
          $ref: '#/components/responses/422ValidationError'
    delete:
      operationId: deactivateCurrentUser
      summary: Деактивировать аккаунт
      description: Мягкое удаление — аккаунт помечается is_active=false.
      tags: [Users]
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Аккаунт деактивирован
        '401':
          $ref: '#/components/responses/401Unauthorized'
...
  /users/me/api-keys:
    get:
      operationId: listApiKeys
      summary: Список API ключей
      tags: [API Keys]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CursorParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Список ключей
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/ApiKeyResponse'
        '401':
          $ref: '#/components/responses/401Unauthorized'
    post:
      operationId: createApiKey
      summary: Создать API ключ
      description: Возвращает полный ключ ТОЛЬКО при создании. Сохраните его.
      tags: [API Keys]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApiKeyCreate'
      responses:
        '201':
          description: Ключ создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyCreated'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '422':
          $ref: '#/components/responses/422ValidationError'
  /users/me/api-keys/{key_id}:
    delete:
      operationId: deleteApiKey
      summary: Удалить API ключ
      tags: [API Keys]
      security:
        - bearerAuth: []
      parameters:
        - name: key_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Ключ удалён
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '404':
          $ref: '#/components/responses/404NotFound'
  # ==========================================================================
  # CHANNELS
  # ==========================================================================
...